implement avro schema

+++ inbound-processor/src/main/java/com/ivan/inbound/enumeration/OrderType.java
@@ -2,22 +2,22 @@ package com.ivan.inbound.enumeration;
 
 public enum OrderType {
 
-    ELECTRONICS("electronics", (byte) 1),
-    CLOTHING("clothing", (byte) 2),
-    FOOD("food", (byte) 3),
-    FURNITURE("furniture", (byte) 4),
-    SPORTS("sports", (byte) 5),
-    OTHER("other", (byte) 0);
+    ELECTRONICS("electronics", 1),
+    CLOTHING("clothing", 2),
+    FOOD("food", 3),
+    FURNITURE("furniture", 4),
+    SPORTS("sports", 5),
+    OTHER("other", 0);
 
     public final String code;
-    public final byte byteCode;
+    public final int byteCode;
 
-    OrderType(String code, byte byteCode) {
+    OrderType(String code, int byteCode) {
         this.code = code;
         this.byteCode = byteCode;
     }
 
-    public static byte fromCode(String code) {
+    public static int fromCode(String code) {
         for (OrderType orderType : OrderType.values()) {
             if (orderType.code.equals(code)) {
                 return orderType.byteCode;
